+++ go.mod
@@ -8,6 +8,7 @@ require (
 	github.com/gorilla/websocket v1.5.3
 	github.com/joho/godotenv v1.5.1
 	github.com/minio/minio-go/v7 v7.0.78
+	github.com/sirupsen/logrus v1.9.3
 	github.com/spf13/viper v1.19.0
 )
