+++ internal/llm/llm.go
@@ -1,0 +1,16 @@
+package llm
+
+type Service interface {
+    GetResponse(prompt string) (string, error)
+}
+
+func NewService(provider string) Service {
+    switch provider {
+    case "openai":
+        return NewOpenAILLM()
+    case "local":
+        return NewLocalLLM()
+    default:
+        return NewLocalLLM()
+    }
+}
